requirements.txt
# Core dependencies for TeleoperatorPC teleoperation
draccus
pyzmq
scservo-sdk
pyserial